+++ src/pages/blog/[slug].astro
@@ -7,7 +7,7 @@ import {
 } from "@/utils/api/blog";
 import BlogcontentWrapper from "@/components/blog/BlogcontentWrapper.vue";
 import BlogDetailsHeroSection from "@/components/blog/BlogDetailsHeroSection.vue";
-import BlogSection from "@/components/contactus/BlogSection.vue";
+import BlogSection from "@/components/blog/BlogSection.vue";
 import { BlogResources } from "@/utils/constant";
 export async function getStaticPaths() {
   const slugs = (await getAllBlogs()).filter((blog) => !blog.caseStudies);
